embedded_storage

Module nor_flash

Technology specific traits for NOR Flashes

Structs

RmwMultiwriteNorFlashStorage

RmwNorFlashStorage

Enums

NorFlashErrorKind

NOR flash error kinds.

Traits

ErrorType

A trait that NorFlash implementations can use to share an error type.

MultiwriteNorFlash

Marker trait for NorFlash relaxing the restrictions on write.

NorFlash

NOR flash trait.

NorFlashError

NOR flash errors.

ReadNorFlash

Read only NOR flash trait.

Functions

check_erase

Return whether an erase operation is aligned and within bounds.

check_read

Return whether a read operation is within bounds.

check_write

Return whether a write operation is aligned and within bounds.