embedded_storage::nor_flash

Trait NorFlashError

pub trait NorFlashError: Debug {
    // Required method
    fn kind(&self) -> NorFlashErrorKind;
}

NOR flash errors.

NOR flash implementations must use an error type implementing this trait. This permits generic code to extract a generic error kind.

Required Methods

fn kind(&self) -> NorFlashErrorKind

Convert a specific NOR flash error into a generic error kind.