embedded_storage::nor_flash
Trait ReadNorFlash

pub trait ReadNorFlash: ErrorType {
    const READ_SIZE: usize;

    // Required methods
    fn read(&mut self, offset: u32, bytes: &mut [u8]) -> Result<(), Self::Error>;
    fn capacity(&self) -> usize;
}

Read only NOR flash trait.

Required Associated Constants

const READ_SIZE: usize

The minumum number of bytes the storage peripheral can read

Required Methods

fn read(&mut self, offset: u32, bytes: &mut [u8]) -> Result<(), Self::Error>

Read a slice of data from the storage peripheral, starting the read operation at the given address offset, and reading bytes.len() bytes.

Errors

Returns an error if the arguments are not aligned or out of bounds. The implementation can use the check_read helper function.

fn capacity(&self) -> usize

The capacity of the peripheral in bytes.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.