embedded_storage

Trait Region

pub trait Region {
    // Required method
    fn contains(&self, address: u32) -> bool;
}

A region denotes a contiguous piece of memory between two addresses.

Required Methods

fn contains(&self, address: u32) -> bool

Check if address is contained in the region of Self