Crate embedded_storage

embedded-storage - A Storage Abstraction Layer for Embedded Systems

Storage traits to allow on and off board storage devices to read and write data.

Modules

iter

Currently contains [OverlapIterator]

nor_flash

Technology specific traits for NOR Flashes

Traits

ReadStorage

Transparent read only storage trait

Region

A region denotes a contiguous piece of memory between two addresses.

Storage

Transparent read/write storage trait