embedded_storage::iter

Trait IterableByOverlaps

pub trait IterableByOverlaps<'a, R, I>
where
    R: Region,
    I: Iterator<Item = R>,
{
    // Required method
    fn overlaps(
        self,
        memory: &'a [u8],
        base_address: u32,
    ) -> OverlapIterator<'_, R, I> ;
}

Trait allowing us to automatically add an overlaps function to all iterators over Region

Required Methods

fn overlaps( self, memory: &'a [u8], base_address: u32, ) -> OverlapIterator<'_, R, I>

Obtain an OverlapIterator over a subslice of memory that overlaps with the region in self