Trait embedded_storage::ReadStorage

pub trait ReadStorage {
    type Error;

    fn read(&mut self, offset: u32, bytes: &mut [u8]) -> Result<(), Self::Error>;
    fn capacity(&self) -> usize;
}

Transparent read only storage trait

Required Associated Types

type Error

An enumeration of storage errors

Required Methods

fn read(&mut self, offset: u32, bytes: &mut [u8]) -> Result<(), Self::Error>

Read a slice of data from the storage peripheral, starting the read operation at the given address offset, and reading bytes.len() bytes.

This should throw an error in case bytes.len() will be larger than self.capacity() - offset.

fn capacity(&self) -> usize

The capacity of the storage peripheral in bytes.