embedded_storage

Trait Storage

pub trait Storage: ReadStorage {
    // Required method
    fn write(&mut self, offset: u32, bytes: &[u8]) -> Result<(), Self::Error>;
}

Transparent read/write storage trait

Required Methods

fn write(&mut self, offset: u32, bytes: &[u8]) -> Result<(), Self::Error>

Write a slice of data to the storage peripheral, starting the write operation at the given address offset (between 0 and self.capacity()).

NOTE: This function will automatically erase any pages necessary to write the given data, and might as such do RMW operations at an undesirable performance impact.